refactor(ui): Move `TimelineStream` into the new `subscriber` module.
This patch moves the `TimelineStream` type into the new `subscriber`
module. The idea is to add more code related to subscribers in the next
patches.
